pyproject.toml
[project]
name = "synology-mcp"
version = "0.1.0"
description = "Comprehensive MCP server for Synology NAS management — file operations, downloads, backups, Docker, photos, system monitoring, and more."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Dima", email = "drivebossllc@gmail.com"}
]
keywords = ["mcp", "synology", "nas", "filestation", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "mcp[cli]>=1.0.0",
    "synology-api>=0.8.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/synology_mcp"]

[project.scripts]
synology-mcp = "synology_mcp.server:main"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
